Rename ide::link_rewrite -> ide::doc_links & tidy imports
