+++ crates/ide/src/doc_links.rs
